django/requirement.txt
Django==3.2.6
djangorestframework-jwt==1.11.0
djangorestframework==3.12.4
drf_yasg==1.20.0
pyyaml==5.4.1
django-filter==2.4.0
mysqlclient==2.1.1
daphne==4.0.0
requests==2.28.2